+++ settings.gradle.kts
@@ -26,7 +26,7 @@ include(":dlt-parser")
 include(":model-contract")
 include(":logger")
 include(":ui-components")
-include(":graphs")
+include(":charts")
 include(":plugins:contract")
 include(":plugins:manager")
 include(":plugins:predefined-plugins")
